minor-fix

+++ src/CryptoPunks.MoreTech.CryptoClient/Models/Requests/RequestTransferNft.cs
@@ -13,5 +13,5 @@ public class RequestTransferNft
 
     public string ToPublicKey { get; set; }
 
-    public int TokenId { get; set; }
+    public long TokenId { get; set; }
 }
