+++ src/CryptoPunks.MoreTech.CryptoClient/Models/Responses/ResponseBalanceNft.cs
@@ -4,5 +4,5 @@ public class ResponseBalanceNft
 {
     public string Url { get; set; } = null!;
 
-    public IEnumerable<int> Tokens { get; set; } = null!;
+    public IEnumerable<long> Tokens { get; set; } = null!;
 }
