+++ src/CryptoPunks.MoreTech.CryptoClient/Models/Responses/ResponseHistoryNft.cs
@@ -2,5 +2,5 @@ namespace CryptoPunks.MoreTech.CryptoClient.Models.Responses;
 
 public class ResponseHistoryNft : ResponseHistoryItem
 {
-    public int TokenId { get; set; }
+    public long TokenId { get; set; }
 }
